requirements.txt
bcrypt==3.1.3
boto3==1.4.2
botocore==1.4.85
cachetools==2.0.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
Django==1.11.2
djangorestframework==3.6.3
docutils==0.13.1
facebook-sdk==2.0.0
firebase-admin==2.0.0
google-auth==1.0.1
idna==2.5
jmespath==0.9.0
mysqlclient==1.3.10
numpy==1.12.0
pyasn1==0.2.3
pyasn1-modules==0.0.9
pycparser==2.17
PyJWT==1.4.2
python-dateutil==2.6.0
pytz==2017.2
requests==2.17.3
rsa==3.4.2
s3transfer==0.1.9
six==1.10.0
urllib3==1.21.1
